backend/image-detect-service/requirements.txt
transformers==4.44.0
requests~=2.28.0
pillow~=11.0.0
kafka-python~=2.2.11
onnxruntime-gpu~=1.22.0
pandas~=2.3.0
numpy~=2.3.0
huggingface-hub~=0.33.0
opencv-python~=4.11.0.86